% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{An example function}
\usage{
example_function(optional_argument = 42L)
}
\arguments{
\item{optional_argument}{an argument with a default value}
}
\description{
An example function
}
\examples{
# Some example code:
df <- example_function()
df

}
