% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_weather.R
\name{fetch_weather}
\alias{fetch_weather}
\title{Fetch standard weather parameters from OpenMeteo for a given latitude/longitude.}
\usage{
fetch_weather(
  latitude,
  longitude,
  year,
  elevation = NA_real_,
  start_date = as_date(str_c(year, "-01-01")),
  end_date = as_date(str_c(year, "12-31")),
  format = TRUE
)
}
\arguments{
\item{latitude}{a scalar latitude (WGS84)}

\item{longitude}{a scalar longitude (WGS84)}

\item{year}{a single year to fetch (ignored if start_date/end_date are changed)}

\item{elevation}{the elevation to use for scraping (optional; if missing the elevation of the point location will be used)}

\item{start_date}{start date for the relevant period (inclusive)}

\item{end_date}{end date for the relevant period (inclusive)}

\item{format}{should the weather data be formatted using \code{\link{format_weather}} before returning?}
}
\value{
a list of data frames giving (1) daily variables, (2) hourly variables (in UTC) for the provided latitude/longitude/elevation
}
\description{
Fetch standard weather parameters from OpenMeteo for a given latitude/longitude.
}
\details{
Fetching data for a 12-month period counts as 60 API calls, so you must not
exceed 10 locations per minute and/or 80 per hour and/or 160 per day (for a
full year at a time). Modifying the time period has a linear effect on the
number of API calls.
}
