% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_weather.R
\name{format_weather}
\alias{format_weather}
\title{Format standard weather parameters from a JSON representation of OpenMeteo data}
\usage{
format_weather(weather)
}
\arguments{
\item{weather}{a JSON representation of OpenMeteo data, obtained from \code{\link{fetch_weather}}}
}
\value{
a nested tibble with daily variables in columns, and tibbles of hourly variables in a list column (see details)
}
\description{
Format standard weather parameters from a JSON representation of OpenMeteo data
}
\details{
Daily variables have the following units:
\itemize{
\item date: R formatted date
\item weather_code: wmo code (TODO: lookup table)
\item sunrise: R formatted date/time (UTC)
\item sunset: R formatted date/time (UTC)
}

Hourly variables have the following units:
\itemize{
\item time: R formatted date/time (UTC)
\item temperature_2m: °C
\item relative_humidity_2m: \%
\item dew_point_2m: °C
\item apparent_temperature: °C
\item precipitation: mm
\item snow_depth: m
\item weather_code: wmo code
\item surface_pressure: hPa
\item wind_speed_10m: km/h
\item wind_speed_100m: km/h
\item wind_direction_10m: °
\item wind_direction_100m: °
\item wind_gusts_10m: km/h
\item soil_temperature_0_to_7cm: °C
\item soil_moisture_0_to_7cm: m³/m³
\item sunshine_duration: s
\item direct_radiation: W/m²
\item et0_fao_evapotranspiration: mm
\item terrestrial_radiation: W/m²
}
}
