% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_to_grid.R
\name{locations_to_grid}
\alias{locations_to_grid}
\title{Convert a data frame of locations (with Latitude and Longitude) into grid IDs}
\usage{
locations_to_grid(locations, grid_size = c("10km", "100km"))
}
\arguments{
\item{locations}{a data frame or tibble including columns Latitude and Longitude}

\item{grid_size}{either 10km or 100km}
}
\description{
Convert a data frame of locations (with Latitude and Longitude) into grid IDs
}
