% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_credentials.R
\name{save_credentials}
\alias{save_credentials}
\title{Title}
\usage{
save_credentials(
  credentials,
  historical = FALSE,
  current = FALSE,
  forecast = FALSE
)
}
\arguments{
\item{forecast}{}
}
\description{
Title
}
