% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_locations.R
\docType{data}
\name{weather_locations}
\alias{weather_locations}
\title{Weather locations for scraping}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1617 rows and 6 columns.
}
\usage{
weather_locations
}
\description{
Weather locations for scraping
}
\details{
A data frame of locations for scraping
}
\keyword{datasets}
