% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherscrape-package.R
\docType{package}
\name{weatherscrape-package}
\alias{weatherscrape-package}
\alias{weatherscrape}
\title{What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\examples{
## To cite this package in publications use:
citation("weatherscrape")

}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[neatpkg]{pkg_new}} for the function that created this package
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
