% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_weather.R
\name{fetch_weather}
\alias{fetch_weather}
\title{Fetch standard weather parameters from OpenMeteo for a given latitude/longitude.}
\usage{
fetch_weather(
  latitude,
  longitude,
  elevation = NA_real_,
  start_date,
  end_date,
  format = TRUE
)
}
\arguments{
\item{latitude}{a scalar latitude (WGS84)}

\item{longitude}{a scalar longitude (WGS84)}

\item{elevation}{the elevation to use for scraping (optional; if missing the elevation of the point location will be used)}

\item{start_date}{start date for the relevant period (inclusive)}

\item{end_date}{end date for the relevant period (inclusive)}

\item{format}{should the weather data be formatted using \code{\link{format_weather}} before returning?}
}
\value{
a list of data frames giving (1) daily variables, (2) hourly variables (in UTC) for the provided latitude/longitude/elevation
}
\description{
Fetch standard weather parameters from OpenMeteo for a given latitude/longitude.
}
\details{
NOTE:  IF YOU USE THIS FUNCTION FROM INSIDE A KU NETWORK OR WHILE
CONNECTED TO THE KU VPN THEN DO SO A MAX OF ONCE PER DAY!
THIS ALSO APPLIES TO YOUR OWN NETWORK IF YOU ARE ALSO USING
THE \code{\link{scrape_weather}} functions

Fetching data for a 12-month period counts as ~60 (2024=57.4) API calls, so you must not
exceed 10 locations per minute and/or 80 per hour and/or 160 per day (for a
full year at a time) - see https://open-meteo.com/en/terms.
Modifying the time period has a linear effect on the number of API calls.
}
