% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_elevation.R
\name{sample_elevation}
\alias{sample_elevation}
\title{Sample elevation for a specified polygon}
\usage{
sample_elevation(grid, n_points = 100L)
}
\arguments{
\item{n_points}{}
}
\description{
Sample elevation for a specified polygon
}
\details{
One call to this function takes a single API call (I think)
}
