% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_elevation.R
\name{sample_elevation}
\alias{sample_elevation}
\title{Sample elevation for a specified polygon}
\usage{
sample_elevation(polygon, n_points = 100L)
}
\arguments{
\item{polygon}{an sf (or sfc) object representing a spatial polygon within which to sample}

\item{n_points}{the number of points to sample}
}
\description{
Sample elevation for a specified polygon
}
\details{
One call to this function takes a single API call (I think)
}
